Link explanation of 'settings' to data tree page
We have made the experience that many users are somewhat familiar with
using dizmoGen but one part that is confusing is how to set dizmo
attributes and which attributes can be set.
This extends the README explaining what "attributes" within "settings"
in the package.json does. It also links to the full documentation of all
the attributes.